Seemingly On-Track For Implementing Jest
- Next: either get TestUtils to work (it's async and not accommodated currently) or use Enzyme (supposedly makes this async handling easy)
