+++ songDisplayComponent/package.json
@@ -10,20 +10,46 @@
   "scripts": {
     "start": "nodemon Server/server.js",
     "build": "webpack -w",
-    "seed": "node db/seedDB.js"
+    "seed": "node db/seedDB.js",
+    "test": "jest --verbose"
   },
   "dependencies": {
+    "axios": "^0.19.0",
     "csv-parse": "^4.6.3",
     "mysql": "^2.17.1",
     "react": "^16.10.2",
     "react-dom": "^16.10.2"
   },
+  "jest": {
+    "modulePaths": [
+      "/shared/vendor/modules"
+    ],
+    "moduleFileExtensions": [
+      "js",
+      "jsx"
+    ],
+    "moduleDirectories": [
+      "node_modules",
+      "bower_components",
+      "shared"
+    ],
+    "moduleNameMapper": {
+      "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js",
+      "\\.(gif|ttf|eot|svg)$": "<rootDir>/__mocks__/fileMock.js"
+    }
+  },
   "devDependencies": {
     "@babel/core": "^7.6.4",
     "@babel/preset-env": "^7.6.3",
     "@babel/preset-react": "^7.6.3",
+    "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
+    "babel-preset-env": "^1.7.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
+    "jest-cli": "^24.9.0",
+    "jest-resolve": "^24.9.0",
     "node-vibrant": "^3.1.4",
     "webpack": "^4.41.1",
     "webpack-cli": "^3.3.9"
